requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
oracledb==3.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
